+++ scripts/make:package:publish:json.js
@@ -1,15 +1,15 @@
 #!/usr/bin/env node
 /* *****************************************************************************
  *
- * Creates the production folder.
+ * Creates package.publish.jon.
  *
- * build:skeleton:prod.js script creates the productiion folder and copies the
- * files defined in config.js.
+ * make:package:publish:json.js script creates the package.json version
+ * to be published on NPM.
  *
  * Private Functions:
  *  . _help                       displays the help message,
- *  . _clean                      removes the previous production build,
- *  . _doskeleton                 creates the production folder and copies files,
+ *  . _clean                      removes the previous package.publish.json,
+ *  . _dopublish                  creates package.publish.json,
  *
  *
  * Public Static Methods:
@@ -23,32 +23,30 @@
  * @since        0.0.0
  * @version      -
  * ************************************************************************** */
-/* eslint curly: 0 */
+/* eslint curly: 1 */
 
 
 // -- Vendor Modules
 import fs from 'fs';
-import path from 'path';
 import nopt from 'nopt';
 
 
 // -- Local Modules
-import config from './config.js';
+import pack from '../package.json' with { type: 'json' };
 
 
 // -- Local Constants
-const VERSION      = '0.0.0-alpha.0'
-    , opts         = {
+const VERSION     = '0.0.0-alpha.0'
+    , opts        = {
       help: [Boolean, false],
       version: [String, null],
     }
-    , shortOpts    = {
+    , shortOpts   = {
       h: ['--help'],
       v: ['--version', VERSION],
     }
-    , parsed       = nopt(opts, shortOpts, process.argv, 2)
-    , { dist }     = config
-    , { webfiles } = config
+    , parsed      = nopt(opts, shortOpts, process.argv, 2)
+    , destination = '.'
     ;
 
 
@@ -70,7 +68,7 @@ function _help() {
   const message = ['',
     'Usage: command [options]',
     '',
-    '                       creates the production folder and copies the files defined in config.js',
+    '                       creates package.publish.json',
     '',
     'Options:',
     '',
@@ -83,7 +81,7 @@ function _help() {
 }
 
 /**
- * Removes the previous production build.
+ * Removes the previous package.publish.json.
  *
  * @function ([arg1])
  * @private
@@ -96,19 +94,23 @@ function _clean(done) {
   process.stdout.write('Starting \'\x1b[36mclean\x1b[89m\x1b[0m\'...\n');
 
   return new Promise((resolve) => {
-    fs.rm(dist, { force: true, recursive: true }, (err) => {
-      if (err) throw new Error(err);
+    fs.rm(`${destination}/package.publish.json`, { force: true }, (err1) => {
+      if (err1) {
+        throw new Error(err1);
+      }
 
       const d2 = new Date() - d1;
       process.stdout.write(`Finished '\x1b[36mclean\x1b[89m\x1b[0m' after \x1b[35m${d2} ms\x1b[89m\x1b[0m\n`);
       resolve();
-      if (done) done();
+      if (done) {
+        done();
+      }
     });
   });
 }
 
 /**
- * Creates the production folder and copies files.
+ * Creates package.publish.json.
  *
  * @function (arg1)
  * @private
@@ -116,31 +118,54 @@ function _clean(done) {
  * @returns {}              -,
  * @since 0.0.0
  */
-function _doskeleton(done) {
+function _dopublish(done) {
   const d1 = new Date();
-  process.stdout.write('Starting \'\x1b[36mdo:skeleton\x1b[89m\x1b[0m\'...\n');
+  process.stdout.write('Starting \'\x1b[36mdo:publish\x1b[89m\x1b[0m\'...\n');
 
-  /**
-   * Wait all processes completed;
-   */
-  let pending = webfiles.length;
-  function _next() {
-    pending -= 1;
-    if (!pending) {
-      const d2 = new Date() - d1;
-      process.stdout.write(`Finished '\x1b[36mdo:skeleton\x1b[89m\x1b[0m' after \x1b[35m${d2} ms\x1b[89m\x1b[0m\n`);
-      done();
+  const pubpack = {};
+  pubpack.name = pack.name;
+  pubpack.version = pack.version;
+  pubpack.description = pack.description;
+  pubpack.main = pack.main;
+  pubpack.minified = pack.minified;
+  pubpack.unpkg = pack.unpkg;
+  pubpack.module = pack.module;
+
+  // pubpack.bin = pack.bin;
+  // pubpack.type = pack.type;
+
+  pubpack.scripts = {};
+  pubpack.scripts.postpack = pack.scripts.postpack;
+
+  pubpack.repository = pack.repository;
+  pubpack.keywords = pack.keywords;
+
+  pubpack.author = pack.author;
+  pubpack.author.name = pack.author.name;
+  pubpack.author.email = pack.author.email;
+  pubpack.author.url = pack.author.url;
+
+  pubpack.license = pack.license;
+  pubpack.bugs = pack.bugs;
+  pubpack.bugs.url = pack.bugs.url;
+  pubpack.homepage = pack.homepage;
+
+  pubpack.dependencies = {};
+  pubpack.devDependencies = {};
+
+  // pubpack.c8 = pack.c8;
+  pubpack.publishConfig = pack.publishConfig;
+  pubpack.private = pack.private;
+
+  fs.writeFile(`${destination}/package.publish.json`, JSON.stringify(pubpack, null, 2), { encoding: 'utf8' }, (err) => {
+    if (err) {
+      throw new Error(err);
     }
-  }
 
-  let filename;
-  for (let i = 0; i < webfiles.length; i++) {
-    filename = path.basename(webfiles[i]);
-    fs.cp(webfiles[i], `${dist}/${filename}`, (err) => {
-      if (err) throw new Error(err);
-      _next();
-    });
-  }
+    const d2 = new Date() - d1;
+    process.stdout.write(`Finished '\x1b[36mdo:dopublish\x1b[89m\x1b[0m' after \x1b[35m${d2} ms\x1b[89m\x1b[0m\n`);
+    done();
+  });
 }
 
 
@@ -171,7 +196,7 @@ const Lib = {
     }
 
     const d1 = new Date();
-    process.stdout.write('Starting \'\x1b[36mbuild:skeleton:prod\x1b[89m\x1b[0m\'...\n');
+    process.stdout.write('Starting \'\x1b[36mmake:package:publish:json\x1b[89m\x1b[0m\'...\n');
 
     let pending = PENDING;
     /**
@@ -181,12 +206,12 @@ const Lib = {
       pending -= 1;
       if (!pending) {
         const d2 = new Date() - d1;
-        process.stdout.write(`Finished '\x1b[36mbuild:skeleton:prod\x1b[89m\x1b[0m' after \x1b[35m${d2} ms\x1b[89m\x1b[0m\n`);
+        process.stdout.write(`Finished '\x1b[36mmake:package:publish:json\x1b[89m\x1b[0m' after \x1b[35m${d2} ms\x1b[89m\x1b[0m\n`);
       }
     }
 
     await _clean();
-    _doskeleton(done);
+    _dopublish(done);
   },
 };